move in_water() to common module
