+++ aodntools/timeseries_products/hourly_timeseries.py
@@ -14,8 +14,7 @@ from pkg_resources import resource_filename
 
 from aodntools import __version__
 from aodntools.timeseries_products import aggregated_timeseries as utils
-from aodntools.timeseries_products.common import NoInputFilesError, check_file, get_qc_variable_names
-
+from aodntools.timeseries_products.common import NoInputFilesError, check_file, get_qc_variable_names, in_water
 
 TEMPLATE_JSON = resource_filename(__name__, 'hourly_timeseries_template.json')
 BINNING_METHOD_JSON = resource_filename(__name__, 'binning_method.json')
@@ -64,19 +63,6 @@ def get_parameter_names(nc):
     """
     params = list(set([s.strip('_quality_control') for s in get_qc_variable_names(nc)]) - set(list(nc.coords)))
     return params
-
-
-def in_water(nc):
-    """
-    cut data to in-water only timestamps, dropping resulting NaN.
-
-    :param nc: xarray dataset
-    :return: xarray dataset
-    """
-    time_deployment_start = np.datetime64(nc.attrs['time_deployment_start'][:-1])
-    time_deployment_end = np.datetime64(nc.attrs['time_deployment_end'][:-1])
-    TIME = nc['TIME'][:]
-    return nc.where((TIME >= time_deployment_start) & (TIME <= time_deployment_end), drop=True)
 
 
 def good_data_only(nc, qcflags):
